Fix warnings in renders

+++ src/Layers/xrRenderPC_R1/LightShadows.cpp
@@ -258,7 +258,7 @@ void CLightShadows::calculate	()
 			// Select slot and set viewport
 			int		s_x			=	slot_id%slot_line;
 			int		s_y			=	slot_id/slot_line;
-			D3DVIEWPORT9 VP		=	{s_x*S_size,s_y*S_size,S_size,S_size,0,1 };
+			D3DVIEWPORT9 VP		=	{(unsigned long) s_x*S_size, (unsigned long) s_y*S_size,S_size,S_size,0,1 };
 			CHK_DX					(HW.pDevice->SetViewport(&VP));
 			
 			// Render object-parts
